+++ src/ast2ram/seminaive/ClauseTranslator.cpp
@@ -688,345 +688,25 @@ Own<ram::Condition> ClauseTranslator::getFunctionalDependencies(const ast::Claus
     return ram::toCondition(dependencies);
 }
 
-// need to define some helper types or we will go insane
-struct PlanTuplesCost {
-    PlanTuplesCost(const std::vector<std::size_t>& givenPlan, std::size_t givenTuples, double givenCost)
-            : plan(givenPlan), tuples(givenTuples), cost(givenCost) {}
-
-    std::vector<std::size_t> plan;
-    std::size_t tuples;
-    double cost;
-};
-
 std::vector<ast::Atom*> ClauseTranslator::getAtomOrdering(const ast::Clause& clause) const {
     auto atoms = ast::getBodyLiterals<ast::Atom>(clause);
-    auto constraints = ast::getBodyLiterals<ast::BinaryConstraint>(clause);
 
-    if (atoms.size() <= 1 || !Global::config().has("auto-schedule")) {
-        // no plan
-        const auto& plan = clause.getExecutionPlan();
-        if (plan == nullptr) {
-            return atoms;
-        }
-
-        // check if there's a plan for the current version
-        auto orders = plan->getOrders();
-        if (!contains(orders, version)) {
-            return atoms;
-        }
-
-        // get the imposed order, and change it to start at zero
-        const auto& order = orders.at(version);
-        auto sz = order->getOrder().size();
-        std::vector<std::size_t> newOrder(sz);
-        std::transform(order->getOrder().begin(), order->getOrder().end(), newOrder.begin(),
-                [](std::size_t i) -> std::size_t { return i - 1; });
-        return reorderAtoms(atoms, newOrder);
+    const auto& plan = clause.getExecutionPlan();
+    if (plan == nullptr) {
+        return atoms;
     }
 
-    assert(context.hasAutoSchedulerStats() && "Must have stats in order to auto-schedule!");
-
-    auto* contextPtr = &context;
-    auto getRelationSize = [&contextPtr](bool isRecursive, const ast::QualifiedName& rel,
-                                   const std::vector<std::size_t>& joinColumns,
-                                   const std::map<std::size_t, std::string>& constantsMap) {
-        std::set<std::size_t> joinKeys(joinColumns.begin(), joinColumns.end());
-        for (auto& [k, _] : constantsMap) {
-            joinKeys.insert(k);
-        }
-
-        if (joinKeys.empty() && !isRecursive) {
-            return contextPtr->getRelationSize(rel);
-        }
-
-        std::stringstream ss;
-        ss << joinKeys;
-        std::string attributes = ss.str();
-        attributes[0] = '[';
-        attributes[attributes.size() - 1] = ']';
-
-        std::stringstream cc;
-        cc << constantsMap;
-        std::string constants = cc.str();
-        constants[0] = '[';
-        constants[constants.size() - 1] = ']';
-
-        if (isRecursive) {
-            return contextPtr->getRecursiveUniqueKeys(rel.toString(), attributes, constants);
-        }
-
-        return contextPtr->getNonRecursiveUniqueKeys(rel.toString(), attributes, constants);
-    };
-
-    std::unordered_set<std::size_t> recursiveInCurrentStratum;
-
-    for (auto* a : sccAtoms) {
-        for (std::size_t i = 0; i < atoms.size(); ++i) {
-            if (*atoms[i] == *a) {
-                recursiveInCurrentStratum.insert(i);
-            }
-        }
+    // check if there's a plan for the current version
+    auto orders = plan->getOrders();
+    if (!contains(orders, version)) {
+        return atoms;
     }
 
-    // map variable name to constants if possible
-    std::unordered_map<std::string, ast::Constant*> varToConstant;
-
-    // map variables to necessary variables on other side of the equality
-    // i.e. x = y + z we should map x -> { y, z }
-    std::unordered_map<std::string, std::set<std::string>> varToOtherVars;
-
-    std::unordered_map<std::string, std::pair<std::set<std::string>, std::set<std::string>>> ineqToUpperLower;
-
-    for (auto* constraint : constraints) {
-        auto* lhs = constraint->getLHS();
-        auto* rhs = constraint->getRHS();
-
-        if (isIneqConstraint(constraint->getBaseOperator())) {
-            if (auto* var = as<ast::Variable>(lhs)) {
-                std::set<std::string> otherVars;
-                visit(rhs, [&](const ast::Variable& v) { otherVars.insert(v.getName()); });
-                if (isLessThan(constraint->getBaseOperator()) || isLessEqual(constraint->getBaseOperator())) {
-                    ineqToUpperLower[var->getName()].second = otherVars;
-                }
-                if (isGreaterThan(constraint->getBaseOperator()) ||
-                        isGreaterEqual(constraint->getBaseOperator())) {
-                    ineqToUpperLower[var->getName()].first = otherVars;
-                }
-            }
-
-            if (auto* var = as<ast::Variable>(rhs)) {
-                std::set<std::string> otherVars;
-                visit(lhs, [&](const ast::Variable& v) { otherVars.insert(v.getName()); });
-                if (isLessThan(constraint->getBaseOperator()) || isLessEqual(constraint->getBaseOperator())) {
-                    ineqToUpperLower[var->getName()].first = otherVars;
-                }
-                if (isGreaterThan(constraint->getBaseOperator()) ||
-                        isGreaterEqual(constraint->getBaseOperator())) {
-                    ineqToUpperLower[var->getName()].second = otherVars;
-                }
-            }
-        }
-
-        // only consider = constraint
-        if (!isEqConstraint(constraint->getBaseOperator())) {
-            continue;
-        }
-
-        if (isA<ast::Variable>(lhs) && isA<ast::Constant>(rhs)) {
-            varToConstant[as<ast::Variable>(lhs)->getName()] = as<ast::Constant>(rhs);
-            continue;
-        }
-
-        if (isA<ast::Constant>(lhs) && isA<ast::Variable>(rhs)) {
-            varToConstant[as<ast::Variable>(rhs)->getName()] = as<ast::Constant>(lhs);
-            continue;
-        }
-
-        if (auto* var = as<ast::Variable>(lhs)) {
-            std::set<std::string> otherVars;
-            visit(rhs, [&](const ast::Variable& v) { otherVars.insert(v.getName()); });
-            varToOtherVars[var->getName()] = otherVars;
-            continue;
-        }
-
-        if (auto* var = as<ast::Variable>(rhs)) {
-            std::set<std::string> otherVars;
-            visit(lhs, [&](const ast::Variable& v) { otherVars.insert(v.getName()); });
-            varToOtherVars[var->getName()] = otherVars;
-            continue;
-        }
-    }
-
-    // check for bounded inequality i.e. EA < EA2 < EA + Size
-    for (auto& p : ineqToUpperLower) {
-        // consider this like an equality
-        auto& [lower, upper] = p.second;
-        if (!lower.empty() && !upper.empty() &&
-                std::includes(upper.begin(), upper.end(), lower.begin(), lower.end())) {
-            varToOtherVars[p.first] = upper;
-        }
-    }
-
-    std::unordered_map<std::size_t, std::set<std::string>> atomIdxToGroundedVars;
-    for (std::size_t i = 0; i < atoms.size(); ++i) {
-        std::set<std::string> groundedVars;
-        visit(*atoms[i], [&](const ast::Variable& v) { groundedVars.insert(v.getName()); });
-        atomIdxToGroundedVars[i] = groundedVars;
-    }
-
-    // #atoms -> variables to join -> plan, cost
-    std::map<std::size_t, std::map<std::set<std::size_t>, PlanTuplesCost>> cache;
-
-    std::unordered_map<std::size_t, std::map<std::size_t, std::string>> atomToIdxConstants;
-
-    std::size_t atomIdx = 0;
-    for (auto* atom : atoms) {
-        std::string name = getClauseAtomName(clause, atom);
-        std::map<std::size_t, std::string> idxConstant;
-
-        std::size_t i = 0;
-        for (auto* argument : atom->getArguments()) {
-            // if we have a variable and a constraint of the form x = 2 then treat x as 2
-            if (auto* var = as<ast::Variable>(argument)) {
-                if (varToConstant.count(var->getName())) {
-                    argument = varToConstant[var->getName()];
-                }
-            }
-
-            if (auto* constant = as<ast::Constant>(argument)) {
-                std::stringstream ss;
-                ss << *translateConstant(*constant);
-                std::string constantValue = ss.str();
-                idxConstant[i] = constantValue;
-            }
-            ++i;
-        }
-
-        atomToIdxConstants[atomIdx] = idxConstant;
-
-        // start by storing the access cost for each individual relation
-        std::vector<std::size_t> empty;
-        bool isRecursive = recursiveInCurrentStratum.count(atomIdx) > 0;
-        std::size_t tuples = getRelationSize(isRecursive, name, empty, idxConstant);
-        double cost = static_cast<double>(tuples * atom->getArity());
-        std::set<std::size_t> singleton = {atomIdx};
-        std::vector<std::size_t> plan = {atomIdx};
-        cache[1].insert(std::make_pair(singleton, PlanTuplesCost(plan, tuples, cost)));
-        ++atomIdx;
-    }
-
-    // do selinger's algorithm
-    std::size_t N = atoms.size();
-    for (std::size_t K = 2; K <= N; ++K) {
-        // for each K sized subset
-        for (auto& subset : context.getSubsets(N, K)) {
-            // remove an entry from the subset
-            for (std::size_t i = 0; i < subset.size(); ++i) {
-                // construct the set S \ S[i]
-                std::set<std::size_t> smallerSubset;
-                for (std::size_t j = 0; j < subset.size(); ++j) {
-                    if (i == j) {
-                        continue;
-                    }
-                    smallerSubset.insert(subset[j]);
-                }
-
-                // lookup the cost in the cache
-                auto& planTuplesCost = cache[K - 1].at(smallerSubset);
-                auto& oldPlan = planTuplesCost.plan;
-                auto oldTuples = planTuplesCost.tuples;
-                auto oldCost = planTuplesCost.cost;
-
-                // compute the grounded variables from the subset
-                std::set<std::string> groundedVariablesFromSubset;
-                for (auto idx : smallerSubset) {
-                    auto& varsGroundedByAtom = atomIdxToGroundedVars[idx];
-                    groundedVariablesFromSubset.insert(varsGroundedByAtom.begin(), varsGroundedByAtom.end());
-                }
-
-                // compute new cost
-                std::size_t atomIdx = subset[i];
-                auto* atom = atoms[atomIdx];
-                std::vector<std::size_t> joinColumns;
-                const auto& args = atom->getArguments();
-                std::size_t numBound = 0;
-                for (std::size_t argIdx = 0; argIdx < args.size(); ++argIdx) {
-                    auto* arg = args[argIdx];
-                    // if we have a constant or var = constant then we ignore
-                    if (atomToIdxConstants[atomIdx].count(argIdx) > 0) {
-                        ++numBound;
-                        continue;
-                    }
-
-                    // unnamed variable i.e. _
-                    if (isA<ast::UnnamedVariable>(arg)) {
-                        ++numBound;
-                        continue;
-                    }
-
-                    if (auto* var = as<ast::Variable>(arg)) {
-                        // free variable so we can't join on it
-                        if (varToOtherVars.count(var->getName()) > 0) {
-                            auto& dependentVars = varToOtherVars.at(var->getName());
-                            if (std::includes(groundedVariablesFromSubset.begin(),
-                                        groundedVariablesFromSubset.end(), dependentVars.begin(),
-                                        dependentVars.end())) {
-                                joinColumns.push_back(argIdx);
-                                ++numBound;
-                                continue;
-                            }
-                        }
-
-                        // direct match on variable
-                        if (groundedVariablesFromSubset.count(var->getName()) > 0) {
-                            joinColumns.push_back(argIdx);
-                            ++numBound;
-                            continue;
-                        }
-                    }
-                }
-
-                bool isRecursive = recursiveInCurrentStratum.count(atomIdx) > 0;
-                std::vector<std::size_t> empty;
-                double expectedTuples = 0;
-
-                if (numBound == atom->getArity()) {
-                    expectedTuples = 1;
-                } else {
-                    auto relSizeWithConstants = getRelationSize(
-                            isRecursive, getClauseAtomName(clause, atom), empty, atomToIdxConstants[atomIdx]);
-
-                    if (joinColumns.empty()) {
-                        expectedTuples = static_cast<double>(relSizeWithConstants);
-                    } else {
-                        auto uniqueKeys = getRelationSize(isRecursive, getClauseAtomName(clause, atom),
-                                joinColumns, atomToIdxConstants[atomIdx]);
-
-                        bool normalize = (uniqueKeys > 0);
-                        expectedTuples =
-                                static_cast<double>(relSizeWithConstants) / (normalize ? uniqueKeys : 1);
-
-                        std::vector<std::size_t> dummy;
-                        for (auto x : oldPlan) {
-                            dummy.push_back(x);
-                        }
-                    }
-                }
-
-                // calculate new number of tuples
-                std::size_t newTuples = static_cast<std::size_t>(oldTuples * expectedTuples);
-
-                // calculate new cost
-                double newCost = oldCost + newTuples * atom->getArity();
-
-                // calculate new plan
-                std::vector<std::size_t> newPlan(oldPlan.begin(), oldPlan.end());
-                newPlan.push_back(atomIdx);
-
-                // if no plan then insert it
-                std::set<std::size_t> currentSet(subset.begin(), subset.end());
-                if (cache[K].count(currentSet) == 0) {
-                    cache[K].insert(std::make_pair(currentSet, PlanTuplesCost(newPlan, newTuples, newCost)));
-                }
-                // if we have a lower cost
-                else if (cache[K].at(currentSet).cost >= newCost) {
-                    cache[K].erase(currentSet);
-                    cache[K].insert(std::make_pair(currentSet, PlanTuplesCost(newPlan, newTuples, newCost)));
-                }
-            }
-        }
-    }
-
-    auto* unsafeClause = const_cast<ast::Clause*>(&clause);
-    unsafeClause->clearExecutionPlan();
-    std::vector<std::size_t> newOrder;
-    assert(cache[N].size() == 1);
-    auto& bestPlanTuplesCost = cache[N].begin()->second;
-    auto& bestPlan = bestPlanTuplesCost.plan;
-    for (std::size_t elem : bestPlan) {
-        newOrder.push_back(elem);
-    }
-
+    // get the imposed order, and change it to start at zero
+    const auto& order = orders.at(version);
+    std::vector<std::size_t> newOrder(order->getOrder().size());
+    std::transform(order->getOrder().begin(), order->getOrder().end(), newOrder.begin(),
+            [](std::size_t i) -> std::size_t { return i - 1; });
     return reorderAtoms(atoms, newOrder);
 }
